+++ pandas/io/formats/latex_format.py
